requirements.txt
streamlit
spacy
PyMuPDF
python-docx
